Rename project, refactor package name
